[Fixed] move `test_timestamp` from ovmf test to example. fixed other nits.
