+++ uefi/src/proto/misc.rs
@@ -37,7 +37,8 @@ impl ResetNotification {
     /// Register a notification function to be called when ResetSystem() is called.
     ///
     ///
-    /// #example
+    /// # Example
+    ///
     /// ```rust
     /// use log::info;
     /// use uefi::Handle;
@@ -46,7 +47,8 @@ impl ResetNotification {
     /// use uefi_raw::Status;
     /// use uefi_raw::table::runtime;
     ///
-    ///     // value efi_reset_fn is the type of ResetSystemFn, a function pointer
+    ///
+    /// /* value efi_reset_fn is the type of ResetSystemFn, a function pointer*/
     ///     unsafe extern "efiapi" fn efi_reset_fn(
     ///             rt: runtime::ResetType,
     ///             status: Status,
@@ -54,16 +56,17 @@ impl ResetNotification {
     ///             data: *const u8,
     ///     ){
     ///         info!("Inside the event callback");
-    ///         // do what you want
+    ///         info!("do what you want");
     ///     }
     ///
     ///     pub fn test(image: Handle, bt: &BootServices) {
     ///
-    ///     /* get protocol*/
     ///         let mut rn = bt
     ///             .open_protocol_exclusive::<ResetNotification>(image)
     ///             .expect("Failed to open Timestamp protocol");
-    ///         rn.register_reset_notify(Some(efi_reset_fn));
+    ///
+    ///         rn.register_reset_notify(Some(efi_reset_fn))
+    ///             .expect("Failed to register a reset notification function!");
     ///     }
     /// ```
     pub fn register_reset_notify(&mut self, reset_function: Option<ResetSystemFn>) -> Result {
@@ -76,7 +79,3 @@ impl ResetNotification {
         unsafe { (self.0.unregister_reset_notify)(&mut self.0, reset_function) }.to_result()
     }
 }
-
-// !TODO: make a safe FFI for raw function 'ResetSystemFn'
-// copy and edit from uefi-raw/src/table/runtime.rs:84 at commit@6093205c3eb27b2e78be4c003c04d46679bff420
-// pub fn new(&self, rt: ResetType, status: Status, data: Option<&[u8]>) ->  Option<ResetSystemFn>
